requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-taggit==2.1.0
graphviz==0.20
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.1.0
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.11.2
pyparsing==3.0.8
sqlparse==0.4.2
zipp==3.8.0
